+++ CHANGES.md
@@ -1,3 +1,8 @@
+0.1.3 (xxxx-xx-xx)
+------------------
+- Added support for relative RV data between the planet and the host star. Only works
+for single planet systems.
+
 0.1.1 (2019-08-15)
 ------------------
 - Refactored code and added documentation. Bumped requirement